refactor: per-file grade A for non-recall files (CLI tests, edge_concurrent split, langchain dedup)
Behaviour-preserving refactors targeting Codacy grade-A on the six
remaining non-A files outside the ANN/recall path. Net -416 lines in
modified files.

Changes:
- velesdb-cli/license.rs: inline tests → license_tests.rs sibling
- velesdb-cli/main.rs: inline tests → main_tests.rs sibling
- edge_concurrent/mod.rs: node-cascade helpers → cascade.rs submodule
  (lock order unchanged; pub(super) visibility only)
- langchain/_common.py: single source for _payload_to_doc helpers
- langchain/multi_query_ops.py + search_ops.py: remove byte-identical
  duplicates, import from _common

All 26 CLI tests, all 55 edge_concurrent tests, and 195/202 langchain
integration tests pass (7 pre-existing environmental failures).
